Default to windowed mode wo/ GTK, not run from TTY
If GTK is not present on compile and the program is run from TTY, with
WINDOWED:PROMPT, the user will be asked for windowed/fullscreen mode, but if run
from the desktop, nothing would happen.

Now windowed mode will be chosen by default if desktop mode is chosen, unless
the mode is specifically chosen in the DF config.

Also reverted Curses to being required for compile, as the game cannot run
otherwise if started without X11 or a display manager.